@@ -11,49 +11,75 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.0.1",
+    "@angular/animations": "~11.0.5",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "~8.0.1",
-    "@angular/compiler": "~8.0.1",
-    "@angular/core": "~8.0.1",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.0.1",
+    "@angular/common": "~11.0.5",
+    "@angular/compiler": "~11.0.5",
+    "@angular/core": "~11.0.5",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "~11.0.5",
+    "@angular/http": "^7.2.16",
+    "@angular/localize": "~11.0.5",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.0.1",
-    "@angular/platform-browser-dynamic": "~8.0.1",
-    "@angular/router": "~8.0.1",
+    "@angular/platform-browser": "~11.0.5",
+    "@angular/platform-browser-dynamic": "~11.0.5",
+    "@angular/router": "~11.0.5",
     "@auth0/auth0-angular": "^1.2.0",
+    "@chargebee/chargebee-js-angular-wrapper": "^0.3.0",
+    "@ng-bootstrap/ng-bootstrap": "^9.0.1",
+    "@password-generator/package": "^2.1.2",
     "@types/jquery": "^3.5.1",
+    "Buffer": "0.0.0",
+    "bootstrap": "^4.5.0",
+    "chargebee": "^2.6.4",
+    "chargebee-typescript": "^2.0.7",
+    "exceljs": "^4.2.0",
+    "file-saver": "^2.0.5",
     "hammerjs": "^2.0.8",
     "highcharts": "^7.2.2",
     "highcharts-angular": "^2.8.0",
+    "http": "0.0.1-security",
+    "http-module": "0.0.5",
+    "https": "^1.0.0",
     "intl-tel-input": "^17.0.8",
     "jspdf": "^2.1.1",
     "jspdf-autotable": "^3.5.13",
+    "latest": "^0.2.0",
+    "moment": "^2.29.1",
     "ng2-tel-input": "^2.3.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-pagination": "^5.0.0",
+    "os": "^0.1.1",
+    "pdfmake": "^0.1.68",
+    "rxjs": "~6.6.3",
+    "stream": "0.0.2",
+    "timers": "^0.1.1",
+    "tslib": "^1.14.1",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.6",
-    "@angular/cli": "~8.0.4",
-    "@angular/compiler-cli": "~8.0.1",
-    "@angular/language-service": "~8.0.1",
+    "@angular-devkit/build-angular": "^0.1100.6",
+    "@angular/cli": "^11.0.5",
+    "@angular/compiler-cli": "~11.0.5",
+    "@angular/language-service": "~11.0.5",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.9.5",
+    "@types/node": "^14.14.20",
     "codelyzer": "^5.0.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
+    "karma": "^5.2.3",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
+    "protractor": "^7.0.0",
+    "ts-node": "^9.1.1",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "typescript": "^4.0.2"
+  },
+  "browser": {
+    "http": false,
+    "https": false
   }
 }
