@@ -1,8 +1,10 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { FormControl, Validators, FormGroup } from '@angular/forms';
 import { UrlConfigService } from 'src/app/shared/url-config.service';
+import { Router } from '@angular/router';
 import { RMIAPIsService } from '../../shared/rmiapis.service';
 import {MatSnackBar,MatSnackBarHorizontalPosition,MatSnackBarVerticalPosition} from '@angular/material/snack-bar';
+import { ExcelService } from 'src/app/shared/excel.service';
 @Component({
   selector: 'app-add-company',
   templateUrl: './add-company.component.html',
@@ -16,19 +18,24 @@ export class AddCompanyComponent implements OnInit {
   companyname: any;
   createdby: any;
   period:any[]=['Monthly', 'Quarterly', 'Yearly'];
-  statementtype:any[]=['Income Statement','Balance Sheet','Cash Flow Statement','All Statements'];
-  industry:any[]=['Communications', 'Consumer & Retail','Distribution & Logistics', 'Energy & Natural Resources','Entertainment & Media','Financial Institutions & Sponsors', 'Food & Beverage', 'General Services', 'Healthcare','Hospitality','Industrials','Power, Infrastructure & Utilities','Real Estate','Technology','Telecommunications','Transportation']
+  //accessType:any[]=['Public', 'Private'];
+  statementType:any[]=['Private','Public'];
+  //industry:any[]=['Communications', 'Consumer & Retail','Distribution & Logistics', 'Energy & Natural Resources','Entertainment & Media','Financial Institutions & Sponsors', 'Food & Beverage', 'General Services', 'Healthcare','Hospitality','Industrials','Power, Infrastructure & Utilities','Real Estate','Technology','Telecommunications','Transportation']
+  industry:any[]= ['Communication Services', 'Consumer Discretionary', 'Consumer Staples', 'Energy', 'Financials', 'Healthcare', 'Industrials', 'Information Technology', 'Materials', 'Pharmaceuticals', 'Real Estate', 'Utilities']
   inprogress: boolean = false;
   form = new FormGroup({
     file: new FormControl(null, [Validators.required]),
     companyname: new FormControl(null, [Validators.required]),
     period: new FormControl(null, [Validators.required]),
-    statementtype: new FormControl(null, [Validators.required]),
-    industry: new FormControl(null, [Validators.required])
+    statementType: new FormControl(null, [Validators.required]),
+    industry: new FormControl(null, [Validators.required]),
+    // accessType: new FormControl(null, [Validators.required])
   });
   constructor(private RMIAPIsService: RMIAPIsService,
     private UrlConfigService:UrlConfigService,
-    private _snackBar: MatSnackBar
+    private _snackBar: MatSnackBar,
+    private router: Router,
+    public excelService: ExcelService
   ) { }
 
 
@@ -58,11 +65,13 @@ export class AddCompanyComponent implements OnInit {
     }
     var postForm = new FormData();
     postForm.append("file", this.form.value.file);
-    postForm.append("statementtype", this.form.value.statementtype);
+    postForm.append("statementType", this.form.value.statementType);
     postForm.append("companyname", this.form.value.companyname);
     postForm.append("createdby", nickname);
     postForm.append("period", this.form.value.period);
     postForm.append("industry", this.form.value.industry);
+    // postForm.append("accessType", this.form.value.accessType);
+    postForm.append("employer", localStorage.getItem("employer")); //
     console.log(JSON.stringify(postForm));
 
    postForm.forEach((value,key) => {
@@ -75,11 +84,21 @@ export class AddCompanyComponent implements OnInit {
       console.log(res.Result,"res upload")
       if(res.Result == "File Uploaded Successfully"){
       this.inprogress = false;
+      this.router.navigate(['/statement']);
       this._snackBar.openFromComponent(uploadSnackBarStatementComponent, {
         duration: 8000,
         horizontalPosition: this.horizontalPosition,
         verticalPosition: this.verticalPosition
       });
+      }
+      else if(res.Result == "Company Name Exist Already, Try Other"){
+        this.inprogress = false;
+        this._snackBar.openFromComponent(snackBarCompanyNameExists, {
+          duration: 8000,
+          horizontalPosition: this.horizontalPosition,
+          verticalPosition: this.verticalPosition
+        });
+
       }
       else if(res.Result == "Extraction Failed"){
         this.inprogress = false;
@@ -139,3 +158,20 @@ export class uploadSnackBarStatementComponent {}
   `],
 })
 export class uploadFailureSnackBarStatementComponent {}
+
+@Component({
+  selector: 'snackBarCompanyNameExists',
+  templateUrl: 'snackBarCompanyNameExists.html',
+  styles: [`
+    .snackBar{
+      color: #fff;
+    }
+    b{
+      color:#fff !important;
+    }
+    .material-icons{
+      color:lightgreen;
+    }
+  `],
+})
+export class snackBarCompanyNameExists {}
